fixing warnings

+++ src/bitvector.c
@@ -7,7 +7,7 @@ extern bitvector_t *bitvector_t_alloc(uint32_t nBits);
 extern void bitvector_t_free(bitvector_t *bv);
 extern void bitvector_t_cleanHighBits(bitvector_t *bv);
 extern void bitvector_t_widen(bitvector_t *bv, uint32_t nBitsToAdd);
-extern uint64_t hexchar_to_digit(char c);
+extern uint8_t hexchar_to_digit(char c);
 extern bitvector_t *bitvector_t_fromHexString(char *string, uint32_t length);
 extern char *bitvector_t_toHexString(bitvector_t *bv);
 extern void bitvector_t_copyUpdate(bitvector_t *dst, bitvector_t *src);
